feat(component): add CardGroup sub-pattern
