+++ packages/react/src/patterns/sections/CardSection/CardSection.js
@@ -5,18 +5,13 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import React, { useEffect, useRef } from 'react';
-import { ArrowRight20 } from '@carbon/icons-react';
-import { Card } from '../../sub-patterns/Card';
+import { CardGroup } from '../../sub-patterns/CardGroup';
 import classNames from 'classnames';
 import { ContentSection } from '../../sub-patterns/ContentSection';
-import { settings as ddsSettings } from '@carbon/ibmdotcom-utilities';
 import PropTypes from 'prop-types';
-import root from 'window-or-global';
-import { sameHeight } from '@carbon/ibmdotcom-utilities';
+import React from 'react';
 import { settings } from 'carbon-components';
 
-const { stablePrefix } = ddsSettings;
 const { prefix } = settings;
 /**
  * Card section Component
@@ -29,43 +24,6 @@ const { prefix } = settings;
  * @returns {object} JSX Object
  */
 const CardSection = ({ heading, theme, cards, cta, ...otherProps }) => {
-  const containerRef = useRef();
-  useEffect(() => {
-    setCardHeight();
-    root.addEventListener('resize', setCardHeight);
-    return () => root.removeEventListener('resize', setCardHeight);
-  }, []);
-
-  /**
-   * Set the cards to have the same height as the bigger one
-   */
-  const setCardHeight = () => {
-    root.requestAnimationFrame(() => {
-      if (containerRef && containerRef.current) {
-        sameHeight(
-          containerRef.current.getElementsByClassName(
-            `${prefix}--card__heading`
-          ),
-          'md'
-        );
-        sameHeight(
-          containerRef.current.getElementsByClassName(`${prefix}--card__copy`),
-          'md'
-        );
-        sameHeight(
-          containerRef.current.getElementsByClassName(
-            `${prefix}--card__eyebrow`
-          ),
-          'md'
-        );
-        sameHeight(
-          containerRef.current.getElementsByClassName(`${prefix}--card--link`),
-          'md'
-        );
-      }
-    });
-  };
-
   /**
    * sets the class name based on theme type
    *
@@ -82,72 +40,16 @@ const CardSection = ({ heading, theme, cards, cta, ...otherProps }) => {
       heading={heading}
       autoid={otherProps.autoid}
       customClassName={classNames(`${prefix}--card-section`, _setTheme(theme))}>
-      {_renderCards(cards, containerRef, cta)}
+      <CardGroup cards={cards} cta={cta} />
     </ContentSection>
   );
 };
 
-/**
- * Renders the cards based on the CardSection entries
- *
- * @param {Array} cards objects array
- * @param {object} containerRef ref of elements
- * @param {object} cta object
- * @returns {*} CardSection JSX objects
- */
-const _renderCards = (cards, containerRef, cta) => (
-  <div
-    data-autoid={`${stablePrefix}--card-section`}
-    className={`${prefix}--card-section__cards__row ${prefix}--row--condensed`}
-    ref={containerRef}>
-    {cards.map((card, index) => {
-      return (
-        <div
-          key={index}
-          className={`${prefix}--card-section__cards__col`}
-          role="region"
-          aria-label={card.heading}>
-          <Card
-            key={index}
-            customClassName={`${prefix}--card-section__card`}
-            image={card.image}
-            heading={card.heading}
-            eyebrow={card.eyebrow}
-            copy={card.copy}
-            cta={{
-              href: card.cta.href,
-              icon: {
-                src: ArrowRight20,
-              },
-            }}
-            type="link"
-          />
-        </div>
-      );
-    })}
-    {cta && (
-      <div className={`${prefix}--card-section__cards__col`}>
-        <Card
-          inverse={true}
-          heading={cta.heading}
-          cta={{
-            href: cta.cta.href,
-            icon: {
-              src: ArrowRight20,
-            },
-          }}
-          type="link"
-        />
-      </div>
-    )}
-  </div>
-);
-
 CardSection.propTypes = {
   theme: PropTypes.string,
   heading: PropTypes.string.isRequired,
-  cards: PropTypes.arrayOf(PropTypes.shape(Card.propTypes)),
-  cta: PropTypes.oneOfType(PropTypes.shape(Card.propTypes)),
+  cards: PropTypes.arrayOf(PropTypes.shape(CardGroup.propTypes)),
+  cta: PropTypes.oneOfType(PropTypes.shape(CardGroup.propTypes)),
 };
 
 export default CardSection;
