+++ packages/react/src/patterns/sections/CardSection/__stories__/CardSection.stories.js
@@ -1,3 +1,10 @@
+/**
+ * Copyright IBM Corp. 2016, 2018
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
 import './index.scss';
 import {
   boolean,
