+++ packages/react/src/patterns/sections/CardSectionSimple/__stories__/CardSectionSimple.stories.js
@@ -1,3 +1,10 @@
+/**
+ * Copyright IBM Corp. 2016, 2018
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
 import {
   boolean,
   object,
