+++ packages/react/src/patterns/sub-patterns/CardGroup/README.md
@@ -1,0 +1,5 @@
+# Card Group
+
+> The CardGroup sub-pattern is a collection of Card components that can be used
+> in block and group -level patterns. It can also be used as a standalone
+> pattern.
